update mosaic list
